app/pyproject.toml
[tool.poetry]
name = "a-lot-of-bq-subscriptions"
version = "0.1.0"
description = ""
authors = ["shin5ok <shin5ok@55mp.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-pubsub = "^2.13.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
